Moved GradiatedPerformanceBar and PassFailBar to their own files
