+++ src/analysis/retail/druid/feral/modules/spells/Bloodtalons.tsx
@@ -25,7 +25,8 @@ import { TALENTS_DRUID } from 'common/TALENTS';
 import { calculateEffectiveDamage } from 'parser/core/EventCalculateLib';
 import { encodeEventTargetString } from 'parser/shared/modules/Enemies';
 import AbilityTracker from 'parser/shared/modules/AbilityTracker';
-import { GradiatedPerformanceBar, SubSection } from 'interface/guide';
+import { SubSection } from 'interface/guide';
+import GradiatedPerformanceBar from 'interface/guide/shared/GradiatedPerformanceBar';
 
 const BUFFER_MS = 50;
