+++ src/analysis/retail/monk/brewmaster/modules/problems/InvokeNiuzao/index.tsx
@@ -2,7 +2,8 @@ import { formatDuration, formatNumber } from 'common/format';
 import SPELLS from 'common/SPELLS';
 import talents from 'common/TALENTS/monk';
 import { AlertWarning, ControlledExpandable, SpellLink, Tooltip } from 'interface';
-import { GuideProps, PassFailBar, Section, SectionHeader, SubSection } from 'interface/guide';
+import { GuideProps, Section, SectionHeader, SubSection } from 'interface/guide';
+import PassFailBar from 'interface/guide/shared/PassFailBar';
 import InformationIcon from 'interface/icons/Information';
 import { AnyEvent } from 'parser/core/Events';
 import { Info } from 'parser/core/metric';
